+++ pkg/util/util.go
@@ -2,7 +2,7 @@ package util
 
 import "math/big"
 
-func Min(a, b *big.Int) *big.Int {
+func Min(a, b *big.Float) *big.Float {
 	if a.Cmp(b) < 0 {
 		return a
 	}
